+++ Python/Batch/sample3_encrypted_resourcefiles.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 # sample3_encrypted_resourcefiles.py Code Sample
 #
 # Copyright (c) Microsoft Corporation
@@ -67,7 +65,7 @@ def generate_secrets(privatekey_pemfile, pfxfile):
          '-keyout', privatekey_pemfile, '-out', 'cert.pem', '-days', '14',
          '-subj', '/C=US/ST=None/L=None/O=None/CN=Test']
     )
-    # convert pem to pfx for service
+    # convert pem to pfx for Azure Batch service
     # NOTE: generated pfx file has a weak passphrase and is only shown below
     # for sample purposes. Always protect private keys with strong passphrases.
     subprocess.check_call(
@@ -87,8 +85,10 @@ def generate_secrets(privatekey_pemfile, pfxfile):
         stdout=subprocess.PIPE
     )
     sha1_cert_tp = proc.communicate(input=pfxdump)[0]
-    return ''.join(sha1_cert_tp.decode(
-        'ascii').strip().split('=')[1].split(':'))
+    # return just the thumbprint from the above openssl command in lowercase
+    # expected openssl output is in the form: SHA1 Fingerprint=<thumbprint>
+    return ''.join(common.helpers.decode_string(
+        sha1_cert_tp).strip().split('=')[1].split(':')).lower()
 
 
 def encrypt_localfile_to_blob_storage(
@@ -102,7 +102,7 @@ def encrypt_localfile_to_blob_storage(
     :param str localresource: local resource file to encrypt
     :param bool rm_rsakey_pemfile: remove RSA key pem file
     :rtype: tuple
-    :return: (rsa private key as a PFX, sha1 thumbprint)
+    :return: (path to rsa private key as a PFX, sha1 thumbprint)
     """
     # create encryption secrets
     rsakeypem = 'rsakey.pem'
@@ -121,22 +121,24 @@ def encrypt_localfile_to_blob_storage(
 
 
 def add_certificate_to_account(
-        batch_client, pfxfile, sha1_cert_tp, rm_pfxfile=True):
+        batch_client, pfxfile, pfx_password, sha1_cert_tp, rm_pfxfile=True):
     """Adds a certificate to a Batch account.
 
     :param batch_client: The batch client to use.
     :type batch_client: `batchserviceclient.BatchServiceClient`
     :param str pfxfile: pfx file to upload
+    :param str pfx_passphrase: pfx password
     :param str sha1_cert_tp: sha1 thumbprint of pfx
-    :param bool rm_pfxfile: remove PFX file
+    :param bool rm_pfxfile: remove PFX file from local disk
     """
     print('adding pfx cert {} to account'.format(sha1_cert_tp))
-    data = base64.b64encode(open(pfxfile, 'rb').read()).decode('ascii')
+    data = common.helpers.decode_string(
+        base64.b64encode(open(pfxfile, 'rb').read()))
     batch_client.certificate.add(
         certificate=batchmodels.CertificateAddParameter(
             sha1_cert_tp, 'sha1', data,
             certificate_format=batchmodels.CertificateFormat.pfx,
-            password=_PFX_PASSPHRASE)
+            password=pfx_password)
     )
     # remove pfxfile
     if rm_pfxfile:
@@ -151,34 +153,32 @@ def create_pool_and_wait_for_node(
     :param batch_client: The batch client to use.
     :type batch_client: `batchserviceclient.BatchServiceClient`
     :param str pool_id: The id of the pool to create.
+    :param str vm_size: vm size (sku)
+    :param int vm_count: number of vms to allocate
     :param str sha1_cert_tp: sha1 cert thumbprint for cert ref
     """
-    node_agent_skus = batch_client.account.list_node_agent_skus()
+    # pick the latest supported 14.04 sku for UbuntuServer
+    sku_to_use, image_ref_to_use = \
+        common.helpers.select_latest_verified_vm_image_with_node_agent_sku(
+            batch_client, 'Canonical', 'UbuntuServer', '14.04')
 
-    # Pick the latest supported sku for UbuntuServer
-    skus_to_use = [(sku, image_ref) for sku in node_agent_skus
-                   for image_ref in sorted(
-                       sku.verified_image_references,
-                       key=lambda item: item.sku)
-                   if image_ref.publisher == "Canonical" and
-                   image_ref.offer == "UbuntuServer" and
-                   "14.04" in image_ref.sku]
-    sku_to_use, image_ref_to_use = skus_to_use[-1]
-
+    # create start task commands
+    # 1. update repository
+    # 2. install blobxfer pre-requisites
+    # 3. pip install blobxfer python script
     start_task_commands = [
         'apt-get update',
         'apt-get install -y build-essential libssl-dev libffi-dev ' +
         'libpython-dev python-dev python-pip',
         'pip install --upgrade blobxfer'
     ]
-    print(common.helpers.wrap_commands_in_shell('linux', start_task_commands))
 
     # create pool with start task and cert ref with visibility of task
     pool = batchmodels.CloudPool(
         id=pool_id,
         virtual_machine_configuration=batchmodels.VirtualMachineConfiguration(
             image_reference=image_ref_to_use,
-            node_agent_sku_id=sku_to_use.id),
+            node_agent_sku_id=sku_to_use),
         vm_size=vm_size,
         target_dedicated=vm_count,
         start_task=batchmodels.StartTask(
@@ -194,44 +194,22 @@ def create_pool_and_wait_for_node(
     )
     common.helpers.create_pool_if_not_exist(batch_client, pool)
 
-    # get first compute node
-    print('waiting for first compute node of pool {}'.format(pool_id))
-    import time
-    while True:
-        nodes = list(batch_client.compute_node.list(pool_id))
-        if len(nodes) > 0:
-            node = nodes[0]
-            break
-        time.sleep(2)
-    print('first node of pool {}: {}'.format(pool_id, node.id))
-
-    # wait for compute node
-    node = common.helpers.wait_for_node_state(
-        batch_client, pool, node,
-        frozenset(
-            (batchmodels.ComputeNodeState.starttaskfailed,
-             batchmodels.ComputeNodeState.unusable,
-             batchmodels.ComputeNodeState.idle)
-        )
-    )
-
-    # ensure node is idle
-    if node.state != batchmodels.ComputeNodeState.idle:
-        raise RuntimeError('node {} of pool {} in unexpected state: {}'.format(
-            node.id, pool.id, node.state))
-
 
 def submit_job_and_add_task(
         batch_client, block_blob_client, storage_account_name,
         storage_account_key, container, resourcefile, job_id, pool_id,
         sha1_cert_tp):
     """Submits a job to the Azure Batch service and adds
-    a task that runs a python script.
+    a task that decrypts the file stored in Azure Storage.
 
     :param batch_client: The batch client to use.
     :type batch_client: `batchserviceclient.BatchServiceClient`
     :param block_blob_client: The storage block blob client to use.
     :type block_blob_client: `azure.storage.blob.BlockBlobService`
+    :param str storage_account_name: storage account name
+    :param str storage_account_key: storage account key
+    :param str container: blob storage container
+    :param str resourcefile: resource file to add to task
     :param str job_id: The id of the job to create.
     :param str pool_id: The id of the pool to use.
     :param str sha1_cert_tp: sha1 cert thumbprint for cert ref
@@ -247,11 +225,19 @@ def submit_job_and_add_task(
         block_blob_client, container, _RESOURCE_NAME,
         azureblob.BlobPermissions.READ)
 
+    # issue the following commands for the task:
+    # 1. convert pfx installed by the Azure Batch Service to pem
+    # 2. transfer the encrypted blob from Azure Storage to local disk and
+    #    decrypt contents using the private key
+    # 3. output decrypted secret.txt file
+    # Note: certs on Linux Batch Compute Nodes are placed in:
+    # $AZ_BATCH_CERTIFICATES_DIR where the cert itself has a naming convention
+    # of <thumbprint algorithm>-<lowercase thumbprint>.<certificate format>
     task_commands = [
         ('openssl pkcs12 -in $AZ_BATCH_CERTIFICATES_DIR/sha1-{tp}.pfx -out '
          '$AZ_BATCH_CERTIFICATES_DIR/privatekey.pem -nodes -password '
          'file:$AZ_BATCH_CERTIFICATES_DIR/sha1-{tp}.pfx.pw').format(
-             tp=sha1_cert_tp.lower()),
+             tp=sha1_cert_tp),
         ('blobxfer {sa} {cont} . --saskey "{sas}" --download '
          '--remoteresource {rf} --rsaprivatekey '
          '$AZ_BATCH_CERTIFICATES_DIR/privatekey.pem').format(
@@ -289,6 +275,9 @@ def execute_sample(global_config, sample_config):
         'Storage',
         'storageaccountsuffix')
 
+    should_delete_container = sample_config.getboolean(
+        'DEFAULT',
+        'shoulddeletecontainer')
     should_delete_job = sample_config.getboolean(
         'DEFAULT',
         'shoulddeletejob')
@@ -337,7 +326,8 @@ def execute_sample(global_config, sample_config):
             _RESOURCE_TO_ENCRYPT)
 
         # add certificate to account
-        add_certificate_to_account(batch_client, rsapfxfile, sha1_cert_tp)
+        add_certificate_to_account(
+            batch_client, rsapfxfile, _PFX_PASSPHRASE, sha1_cert_tp)
 
         # create pool and wait for node idle
         create_pool_and_wait_for_node(
@@ -363,7 +353,7 @@ def execute_sample(global_config, sample_config):
         common.helpers.wait_for_tasks_to_complete(
             batch_client,
             job_id,
-            datetime.timedelta(minutes=5))
+            datetime.timedelta(minutes=20))
 
         tasks = batch_client.task.list(job_id)
         task_ids = [task.id for task in tasks]
@@ -371,9 +361,11 @@ def execute_sample(global_config, sample_config):
         common.helpers.print_task_output(batch_client, job_id, task_ids)
     finally:
         # perform clean up
-        block_blob_client.delete_container(
-            _CONTAINER_NAME,
-            fail_not_exist=False)
+        if should_delete_container:
+            print('Deleting container: {}'.format(_CONTAINER_NAME))
+            block_blob_client.delete_container(
+                _CONTAINER_NAME,
+                fail_not_exist=False)
         if should_delete_job:
             print('Deleting job: {}'.format(job_id))
             batch_client.job.delete(job_id)
@@ -381,6 +373,14 @@ def execute_sample(global_config, sample_config):
             print('Deleting pool: {}'.format(pool_id))
             batch_client.pool.delete(pool_id)
         if should_delete_cert and sha1_cert_tp is not None:
+            # cert deletion requires no active references to cert, so
+            # override any config settings for preserving job/pool
+            if not should_delete_job:
+                print('Deleting job: {}'.format(job_id))
+                batch_client.job.delete(job_id)
+            if not should_delete_pool:
+                print('Deleting pool: {}'.format(pool_id))
+                batch_client.pool.delete(pool_id)
             print('Deleting cert: {}'.format(sha1_cert_tp))
             batch_client.certificate.delete('sha1', sha1_cert_tp)
